nedxml: fix #300: wrong printf() format for error line number
Note: in older Expat versions (e.g. 1.95.8), XML_GetCurrentLineNumber()
return type was int; at some point this was apparently changed to long.
Added explicit cast to accomodate both.

X-Committed-on-Branch: master

+++ src/nedxml/saxparser_expat.cc
@@ -131,9 +131,9 @@ bool SAXParser::parse(const char *filename)
         done = len < sizeof(buf);
         if (!XML_Parse(parser, buf, len, done))
         {
-            sprintf(errortext, "Parse error: %s at line %d",
+            sprintf(errortext, "Parse error: %s at line %ld",
                     XML_ErrorString(XML_GetErrorCode(parser)),
-                    XML_GetCurrentLineNumber(parser));
+                    (long)XML_GetCurrentLineNumber(parser));
             err = true;
             done = true;
         }
